split Migrate class
