--- src/Migrate.php
@@ -1,1878 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace ProjectMigrationTool;
-
-use Keboola\Component\UserException;
-use Keboola\SnowflakeDbAdapter\Exception\RuntimeException;
-use Keboola\SnowflakeDbAdapter\QueryBuilder;
-use ProjectMigrationTool\Configuration\Config;
-use ProjectMigrationTool\Exception\NoWarehouseException;
-use ProjectMigrationTool\Snowflake\Connection;
-use ProjectMigrationTool\Snowflake\Helper;
-use Psr\Log\LoggerInterface;
-use Symfony\Component\Process\Process;
-use Throwable;
-
-class Migrate
-{
-    private LoggerInterface $logger;
-
-    private Config $config;
-
-    private Connection $sourceConnection;
-
-    private ?Connection $migrateConnection;
-
-    private Connection $destinationConnection;
-
-    private array $databases;
-
-    private string $mainMigrationRoleSourceAccount;
-
-    private string $mainMigrationRoleTargetAccount;
-
-    private const MIGRATION_SHARE_PREFIX = 'MIGRATION_SHARE_';
-
-    private const SKIP_CLONE_SCHEMAS = [
-        'INFORMATION_SCHEMA',
-        'PUBLIC',
-    ];
-
-    private array $usedUsers = [];
-
-    public function __construct(
-        LoggerInterface $logger,
-        Config $config,
-        Connection $sourceConnection,
-        ?Connection $migrateConnection,
-        Connection $destinationConnection,
-        array $databases,
-        string $mainMigrationRoleSourceAccount,
-        string $mainMigrationRoleTargetAccount
-    ) {
-        $this->logger = $logger;
-        $this->config = $config;
-        $this->sourceConnection = $sourceConnection;
-        $this->migrateConnection = $migrateConnection;
-        $this->destinationConnection = $destinationConnection;
-        $this->databases = $databases;
-        $this->mainMigrationRoleSourceAccount = $mainMigrationRoleSourceAccount;
-        $this->mainMigrationRoleTargetAccount = $mainMigrationRoleTargetAccount;
-    }
-
-    public function cleanupAccount(string $mainRoleName, bool $dryRun = true): void
-    {
-        $sqls = [];
-        $currentRole = $this->mainMigrationRoleTargetAccount;
-        $this->destinationConnection->grantRoleToUser($this->config->getTargetSnowflakeUser(), $mainRoleName);
-        foreach ($this->databases as $database) {
-            $this->destinationConnection->useRole($this->mainMigrationRoleTargetAccount);
-
-            $dbExists = $this->destinationConnection->fetchAll(sprintf(
-                'SHOW DATABASES LIKE %s;',
-                QueryBuilder::quote($database)
-            ));
-            if (!$dbExists) {
-                continue;
-            }
-            $databaseRole = $this->destinationConnection->getOwnershipRoleOnDatabase($database);
-            $data = $this->getDataToRemove($databaseRole);
-
-            $currentUser = $this->destinationConnection->fetchAll('SELECT CURRENT_USER() AS "user";');
-            foreach ($data['USER'] ?? [] as $user) {
-                if ($user['granted_by'] !== $currentRole) {
-                    $currentRole = $user['granted_by'];
-                    $sqls[] = sprintf(
-                        'GRANT ROLE %s TO USER %s;',
-                        Helper::quoteIdentifier($currentRole),
-                        Helper::quoteIdentifier($this->config->getTargetSnowflakeUser())
-                    );
-                    $sqls[] = sprintf('USE ROLE %s;', Helper::quoteIdentifier($currentRole));
-                }
-
-                if ($currentUser[0]['user'] !== $user['name']) {
-                    $sqls[] = sprintf('DROP USER IF EXISTS %s;', Helper::quoteIdentifier($user['name']));
-                }
-            }
-
-            foreach ($data['ROLE'] ?? [] as $role) {
-                if ($role['granted_by'] !== $currentRole) {
-                    $currentRole = $role['granted_by'];
-                    $sqls[] = sprintf('USE ROLE %s;', Helper::quoteIdentifier($currentRole));
-                }
-                $this->destinationConnection->useRole($mainRoleName);
-                $this->destinationConnection->query(sprintf(
-                    'GRANT ROLE %s TO USER %s;',
-                    Helper::quoteIdentifier($currentRole),
-                    Helper::quoteIdentifier($this->config->getTargetSnowflakeUser())
-                ));
-                $this->destinationConnection->useRole($role['granted_by']);
-                $futureGrants = $this->destinationConnection->fetchAll(sprintf(
-                    'SHOW FUTURE GRANTS TO ROLE %s',
-                    $role['name']
-                ));
-                foreach ($futureGrants as $futureGrant) {
-                    $sqls[] = sprintf(
-                        'REVOKE %s ON FUTURE TABLES IN SCHEMA %s FROM ROLE %s;',
-                        $futureGrant['privilege'],
-                        Helper::removeStringFromEnd($futureGrant['name'], '.<TABLE>'),
-                        Helper::quoteIdentifier($futureGrant['grantee_name']),
-                    );
-                }
-                $sqls[] = sprintf('DROP ROLE IF EXISTS %s;', Helper::quoteIdentifier($role['name']));
-            }
-
-            $grantsOfRole = $this->destinationConnection->fetchAll(sprintf('SHOW GRANTS OF ROLE %s', $databaseRole));
-            $filteredGrantsOfRole = array_filter(
-                $grantsOfRole,
-                fn($v) => strtoupper($v['grantee_name']) === strtoupper($databaseRole)
-            );
-            assert(count($filteredGrantsOfRole) === 1);
-            $grantOfRole = current($filteredGrantsOfRole);
-            if ($grantOfRole['granted_by'] !== $currentRole) {
-                $currentRole = $grantOfRole['granted_by'];
-                $sqls[] = sprintf(
-                    'GRANT ROLE %s TO USER %s;',
-                    Helper::quoteIdentifier($currentRole),
-                    Helper::quoteIdentifier($this->config->getTargetSnowflakeUser())
-                );
-                $sqls[] = sprintf('USE ROLE %s;', Helper::quoteIdentifier($currentRole));
-            }
-            $sqls[] = sprintf(
-                'DROP USER IF EXISTS %s;',
-                Helper::quoteIdentifier($grantOfRole['grantee_name'])
-            );
-
-            $sqls[] = sprintf('DROP ROLE IF EXISTS %s;', Helper::quoteIdentifier($databaseRole));
-
-            $sqls[] = sprintf(
-                'DROP DATABASE IF EXISTS %s;',
-                Helper::quoteIdentifier($database . '_OLD')
-            );
-
-            $sqls[] = sprintf(
-                'ALTER DATABASE IF EXISTS %s RENAME TO %s;',
-                Helper::quoteIdentifier($database),
-                Helper::quoteIdentifier($database . '_OLD'),
-            );
-        }
-
-        foreach ($sqls as $sql) {
-            if ($dryRun) {
-                $this->logger->info($sql);
-            } else {
-                $this->destinationConnection->query($sql);
-            }
-        }
-        if ($dryRun && $sqls) {
-            throw new UserException('!!! PLEASE RUN SQLS ON TARGET SNOWFLAKE ACCOUNT !!!');
-        }
-
-        $this->destinationConnection->query(sprintf(
-            'USE ROLE %s',
-            Helper::quoteIdentifier($this->mainMigrationRoleTargetAccount)
-        ));
-    }
-
-    public function createReplication(): void
-    {
-        if ($this->sourceConnection->getRegion() === $this->destinationConnection->getRegion()) {
-            return;
-        }
-        if (!$this->migrateConnection) {
-            throw new UserException('Migration connection is not set');
-        }
-        foreach ($this->databases as $database) {
-            //            Allow replication on source database
-            $this->sourceConnection->query(sprintf(
-                'ALTER DATABASE %s ENABLE REPLICATION TO ACCOUNTS %s.%s;',
-                Helper::quoteIdentifier($database),
-                $this->migrateConnection->getRegion(),
-                $this->migrateConnection->getAccount()
-            ));
-            $this->sourceConnection->useRole($this->sourceConnection->getOwnershipRoleOnDatabase($database));
-
-            //            Waiting for previous SQL query
-            sleep(1);
-
-            //            Migration database sqls
-            $this->migrateConnection->query(sprintf(
-                'CREATE DATABASE IF NOT EXISTS %s AS REPLICA OF %s.%s.%s;',
-                Helper::quoteIdentifier($database),
-                $this->sourceConnection->getRegion(),
-                $this->sourceConnection->getAccount(),
-                Helper::quoteIdentifier($database)
-            ));
-
-            $this->migrateConnection->query(sprintf(
-                'USE DATABASE %s',
-                Helper::quoteIdentifier($database)
-            ));
-
-            $this->migrateConnection->query('USE SCHEMA PUBLIC');
-
-            //            Create and use warehouse for replicate data
-            $sql = <<<SQL
-CREATE WAREHOUSE IF NOT EXISTS "migrate"
-    WITH WAREHOUSE_SIZE = 'Small'
-        WAREHOUSE_TYPE = 'STANDARD'
-        AUTO_SUSPEND = 300
-        AUTO_RESUME = true
-;
-SQL;
-            $this->migrateConnection->query($sql);
-
-            $this->migrateConnection->query('USE WAREHOUSE "migrate";');
-
-            //            Run replicate of data
-            $this->migrateConnection->query(sprintf(
-                'ALTER DATABASE %s REFRESH',
-                Helper::quoteIdentifier($database)
-            ));
-        }
-    }
-
-    public function createShare(): void
-    {
-        $sourceRegion = $this->sourceConnection->getRegion();
-        $destinationRegion = $this->destinationConnection->getRegion();
-
-        $connection = $this->sourceConnection;
-        if ($sourceRegion !== $destinationRegion) {
-            if (!$this->migrateConnection) {
-                throw new UserException('Migration connection is not set');
-            }
-            $connection = $this->migrateConnection;
-        }
-
-        foreach ($this->databases as $database) {
-            $shareName = sprintf('%s%s', self::MIGRATION_SHARE_PREFIX, strtoupper($database));
-
-            $connection->query(sprintf(
-                'CREATE SHARE IF NOT EXISTS %s;',
-                Helper::quoteIdentifier($shareName)
-            ));
-
-            $connection->query(sprintf(
-                'GRANT USAGE ON DATABASE %s TO SHARE %s;',
-                Helper::quoteIdentifier($database),
-                Helper::quoteIdentifier($shareName)
-            ));
-
-            $connection->query(sprintf(
-                'GRANT USAGE ON ALL SCHEMAS IN DATABASE %s TO SHARE %s;',
-                Helper::quoteIdentifier($database),
-                Helper::quoteIdentifier($shareName)
-            ));
-
-            $connection->query(sprintf(
-                'GRANT SELECT ON ALL TABLES IN DATABASE %s TO SHARE %s;',
-                Helper::quoteIdentifier($database),
-                Helper::quoteIdentifier($shareName)
-            ));
-
-            $connection->query(sprintf(
-                'ALTER SHARE %s ADD ACCOUNT=%s;',
-                Helper::quoteIdentifier($shareName),
-                $this->destinationConnection->getAccount()
-            ));
-        }
-    }
-
-    public function createDatabasesFromShares(): void
-    {
-        $sourceRegion = $this->sourceConnection->getRegion();
-        $destinationRegion = $this->destinationConnection->getRegion();
-
-        $connection = $this->sourceConnection;
-        if ($sourceRegion !== $destinationRegion) {
-            if (!$this->migrateConnection) {
-                throw new UserException('Migration connection is not set');
-            }
-            $connection = $this->migrateConnection;
-        }
-
-        foreach ($this->databases as $database) {
-            $shareDbName = $database . '_SHARE';
-
-            $this->destinationConnection->query(sprintf(
-                'DROP DATABASE IF EXISTS %s;',
-                Helper::quoteIdentifier($shareDbName)
-            ));
-
-            $this->destinationConnection->query(sprintf(
-                'CREATE DATABASE %s FROM SHARE IDENTIFIER(\'%s.%s\');',
-                Helper::quoteIdentifier($shareDbName),
-                $connection->getAccount(),
-                self::MIGRATION_SHARE_PREFIX . $database
-            ));
-        }
-    }
-
-    public function migrateUsersRolesAndGrants(string $mainRole, array $grants): void
-    {
-        $this->logger->info('Migrating users and roles.');
-        // first step - migrate users and roles (without grants)
-        foreach ($this->databases as $database) {
-            $databaseRole = $this->sourceConnection->getOwnershipRoleOnDatabase($database);
-            [
-                'grants' => [
-                    'roles' => $rolesGrants,
-                    'account' => $accountGrants,
-                    'user' => $userGrants,
-                ],
-            ] = Helper::parseGrantsToObjects($grants[$databaseRole]);
-
-            $this->destinationConnection->createRole([
-                'name' => $databaseRole,
-                'granted_by' => $mainRole,
-                'privilege' => 'OWNERSHIP',
-            ], $this->config->getTargetSnowflakeUser());
-
-            foreach ($accountGrants as $grant) {
-                $this->destinationConnection->assignGrantToRole($grant);
-            }
-
-            foreach ($rolesGrants as $rolesGrant) {
-                if ($rolesGrant['privilege'] === 'OWNERSHIP') {
-                    $this->destinationConnection->createRole($rolesGrant, $this->config->getTargetSnowflakeUser());
-                }
-            }
-
-            foreach ($userGrants as $userGrant) {
-                $this->createUser($userGrant, $this->config->getPasswordOfUsers());
-            }
-        }
-
-        $this->logger->info('Migrating grants of warehouses/users and roles.');
-        // second step - migrate all grants of roles/users/warehouses/account
-        foreach ($this->databases as $database) {
-            $databaseRole = $this->sourceConnection->getOwnershipRoleOnDatabase($database);
-
-            [
-                'grants' => [
-                    'roles' => $rolesGrants,
-                    'warehouse' => $warehouseGrants,
-                    'user' => $userGrants,
-                ],
-            ] = Helper::parseGrantsToObjects($grants[$databaseRole]);
-
-            foreach ($rolesGrants as $rolesGrant) {
-                $this->destinationConnection->assignGrantToRole($rolesGrant);
-            }
-
-            foreach ($warehouseGrants as $warehouseGrant) {
-                $this->destinationConnection->assignGrantToRole($warehouseGrant);
-            }
-
-            foreach ($userGrants as $userGrant) {
-                $this->destinationConnection->assignGrantToRole($userGrant);
-            }
-        }
-    }
-
-    public function reApplyFailedGrants(): void
-    {
-        $failedGrants = $this->destinationConnection->getFailedGrants();
-        if (count($failedGrants) === 0) {
-            $this->logger->info('There were no failed grants, nothing to apply');
-            return;
-        }
-        $this->logger->info('Attempting to apply ' . count($failedGrants) . ' failed grants');
-        foreach ($failedGrants as $grant) {
-            $this->destinationConnection->assignGrantToRole($grant);
-        }
-    }
-
-    public function cloneDatabaseWithGrants(string $mainRole, array $grants): void
-    {
-        foreach ($this->databases as $database) {
-            $this->sourceConnection->useRole($this->mainMigrationRoleSourceAccount);
-            $databaseRole = $this->sourceConnection->getOwnershipRoleOnDatabase($database);
-            [
-                'grants' => [
-                    'databases' => $databaseGrants,
-                    'schemas' => $schemasGrants,
-                    'tables' => $tablesGrants,
-                    'views' => $viewsGrants,
-                    'functions' => $functionsGrants,
-                    'procedures' => $proceduresGrants,
-                ],
-                'futureGrants' => [
-                    'tables' => $tablesFutureGrants,
-                ],
-            ] = Helper::parseGrantsToObjects($grants[$databaseRole]);
-
-            $this->destinationConnection->useRole($mainRole);
-
-            $shareDbName = $database . '_SHARE';
-            $oldDbName = $database . '_OLD';
-
-            $sourceDatabases = $this->sourceConnection->fetchAll(sprintf(
-                'SHOW DATABASES LIKE %s',
-                QueryBuilder::quote($database)
-            ));
-            assert(count($sourceDatabases) === 1);
-            $sourceDatabase = current($sourceDatabases);
-
-            $this->logger->info(sprintf('Migrate database "%s".', $database));
-            $this->destinationConnection->query(sprintf(
-                'CREATE DATABASE %s DATA_RETENTION_TIME_IN_DAYS=%s;',
-                Helper::quoteIdentifier($database),
-                $sourceDatabase['retention_time']
-            ));
-
-            foreach ($databaseGrants as $databaseGrant) {
-                if ($databaseGrant['privilege'] === 'OWNERSHIP') {
-                    $this->destinationConnection->assignGrantToRole(array_merge(
-                        $databaseGrant,
-                        ['granted_by' => $mainRole]
-                    ));
-                }
-                $this->destinationConnection->assignGrantToRole($databaseGrant);
-            }
-
-            $this->assignSharePrivilegesToRole($database, $databaseRole);
-            $this->assignForeignGrants($schemasGrants, $databaseRole);
-
-            $this->destinationConnection->useRole($databaseRole);
-
-            $schemas = $this->destinationConnection->fetchAll(sprintf(
-                'SHOW SCHEMAS IN DATABASE %s;',
-                Helper::quoteIdentifier($shareDbName)
-            ));
-
-            foreach ($schemas as $schema) {
-                if (in_array($schema['name'], self::SKIP_CLONE_SCHEMAS)) {
-                    continue;
-                }
-                $schemaName = $schema['name'];
-
-                $this->logger->info(sprintf('Migrate schema "%s".', $schemaName));
-
-                $schemaGrants = Helper::filterSchemaGrants($database, $schemaName, $schemasGrants);
-                $schemaFutureGrants = Helper::filterSchemaGrants($database, $schemaName, $tablesFutureGrants);
-                $ownershipOnSchema = array_filter($schemaGrants, fn($v) => $v['privilege'] === 'OWNERSHIP');
-                assert(count($ownershipOnSchema) === 1);
-
-                $schemaOptions = array_map(fn($v) => trim($v), explode(',', $schema['options']));
-
-                $this->destinationConnection->useRole(current($ownershipOnSchema)['granted_by']);
-                $this->destinationConnection->query(sprintf(
-                    'CREATE %s SCHEMA %s.%s %s DATA_RETENTION_TIME_IN_DAYS = %s;',
-                    in_array('TRANSIENT', $schemaOptions) ? 'TRANSIENT' : '',
-                    Helper::quoteIdentifier($database),
-                    Helper::quoteIdentifier($schemaName),
-                    in_array('MANAGED ACCESS', $schemaOptions) ? 'WITH MANAGED ACCESS' : '',
-                    $schema['retention_time']
-                ));
-
-                foreach ($schemaGrants as $schemaGrant) {
-                    $this->destinationConnection->assignGrantToRole($schemaGrant);
-                }
-                foreach ($schemaFutureGrants as $schemaFutureGrant) {
-                    $this->destinationConnection->assignFutureGrantToRole($schemaFutureGrant);
-                }
-
-                $tables = $this->destinationConnection->fetchAll(sprintf(
-                    'SHOW TABLES IN SCHEMA %s.%s;',
-                    Helper::quoteIdentifier($shareDbName),
-                    Helper::quoteIdentifier($schemaName)
-                ));
-
-                foreach ($tables as $table) {
-                    $tableName = $table['name'];
-
-                    $tableGrants = array_filter(
-                        $tablesGrants,
-                        function (array $v) use ($database, $schemaName, $tableName) {
-                            $validSchema = [
-                                sprintf('%s.%s.%s', $database, $schemaName, $tableName),
-                                sprintf('%s.%s.%s', $database, $schemaName, Helper::quoteIdentifier($tableName)),
-                                sprintf('%s.%s.%s', $database, Helper::quoteIdentifier($schemaName), $tableName),
-                                sprintf(
-                                    '%s.%s.%s',
-                                    $database,
-                                    Helper::quoteIdentifier($schemaName),
-                                    Helper::quoteIdentifier($tableName)
-                                ),
-                                sprintf(
-                                    '%s.%s.%s',
-                                    Helper::quoteIdentifier($database),
-                                    $schemaName,
-                                    $tableName
-                                ),
-                                sprintf(
-                                    '%s.%s.%s',
-                                    Helper::quoteIdentifier($database),
-                                    $schemaName,
-                                    Helper::quoteIdentifier($tableName)
-                                ),
-                                sprintf(
-                                    '%s.%s.%s',
-                                    Helper::quoteIdentifier($database),
-                                    Helper::quoteIdentifier($schemaName),
-                                    $tableName
-                                ),
-                                sprintf(
-                                    '%s.%s.%s',
-                                    Helper::quoteIdentifier($database),
-                                    Helper::quoteIdentifier($schemaName),
-                                    Helper::quoteIdentifier($tableName)
-                                ),
-                            ];
-                            return in_array($v['name'], $validSchema);
-                        }
-                    );
-                    $ownershipOnTable = array_filter($tableGrants, fn($v) => $v['privilege'] === 'OWNERSHIP');
-                    assert(count($ownershipOnTable) === 1);
-
-                    $ownershipOnTable = current($ownershipOnTable);
-
-                    self::assignSharePrivilegesToRole(
-                        $database,
-                        $ownershipOnTable['granted_by']
-                    );
-                    $this->destinationConnection->useRole($ownershipOnTable['granted_by']);
-                    try {
-                        $this->destinationConnection->useWarehouse($ownershipOnTable['granted_by']);
-                    } catch (NoWarehouseException $exception) {
-                        if (!preg_match('/^(KEBOOLA|SAPI|sapi)_WORKSPACE_/', $ownershipOnTable['granted_by'])) {
-                            throw $exception;
-                        }
-                        $this->logger->info(sprintf(
-                            'Warning: Skipping table: %s, because: %s',
-                            $tableName,
-                            $exception->getMessage()
-                        ));
-                    }
-
-                    if ($this->canCloneTable($mainRole, $database, $schemaName, $tableName)) {
-                        $this->grantUsageToOldTable(
-                            $database,
-                            $schemaName,
-                            $tableName,
-                            $mainRole,
-                            $ownershipOnTable['granted_by']
-                        );
-                        $this->logger->info(sprintf('Cloning table "%s" from OLD database', $tableName));
-                        $this->destinationConnection->query(sprintf(
-                            'CREATE TABLE %s.%s.%s CLONE %s.%s.%s;',
-                            Helper::quoteIdentifier($database),
-                            Helper::quoteIdentifier($schemaName),
-                            Helper::quoteIdentifier($tableName),
-                            Helper::quoteIdentifier($oldDbName),
-                            Helper::quoteIdentifier($schemaName),
-                            Helper::quoteIdentifier($tableName),
-                        ));
-                    } else {
-                        $this->logger->info(sprintf('Creating table "%s" from SHARE database', $tableName));
-
-                        try {
-                            $this->destinationConnection->query(sprintf(
-                                'CREATE TABLE %s.%s.%s LIKE %s.%s.%s;',
-                                Helper::quoteIdentifier($database),
-                                Helper::quoteIdentifier($schemaName),
-                                Helper::quoteIdentifier($tableName),
-                                Helper::quoteIdentifier($shareDbName),
-                                Helper::quoteIdentifier($schemaName),
-                                Helper::quoteIdentifier($tableName),
-                            ));
-
-                            $this->destinationConnection->query(sprintf(
-                                'INSERT INTO %s.%s.%s SELECT * FROM %s.%s.%s;',
-                                Helper::quoteIdentifier($database),
-                                Helper::quoteIdentifier($schemaName),
-                                Helper::quoteIdentifier($tableName),
-                                Helper::quoteIdentifier($shareDbName),
-                                Helper::quoteIdentifier($schemaName),
-                                Helper::quoteIdentifier($tableName),
-                            ));
-                        } catch (RuntimeException $e) {
-                            $this->logger->info(sprintf(
-                                'Warning: Skip creating table %s.%s.%s. Error: "%s".',
-                                Helper::quoteIdentifier($database),
-                                Helper::quoteIdentifier($schemaName),
-                                Helper::quoteIdentifier($tableName),
-                                $e->getMessage()
-                            ));
-                        }
-                    }
-
-                    $tableGrants = array_filter($tableGrants, fn($v) => $v['privilege'] !== 'OWNERSHIP');
-                    foreach ($tableGrants as $tableGrant) {
-                        $this->destinationConnection->assignGrantToRole($tableGrant);
-                    }
-                }
-            }
-
-            $this->copyViews($database, $databaseRole, $viewsGrants);
-
-            $this->copyFunctions($database, $functionsGrants);
-
-            $this->copyProcedures($database, $proceduresGrants);
-        }
-    }
-
-    public function getMainRoleWithGrants(): array
-    {
-        $grantsOfRoles = [];
-        foreach ($this->databases as $database) {
-            $roleName = $this->sourceConnection->getOwnershipRoleOnDatabase($database);
-
-            $grantedOnDatabaseRole = $this->sourceConnection->fetchAll(sprintf(
-                'SHOW GRANTS ON ROLE %s',
-                Helper::quoteIdentifier($roleName)
-            ));
-
-            $ownershipOfRole = array_filter($grantedOnDatabaseRole, fn($v) => $v['privilege'] === 'OWNERSHIP');
-
-            $ownershipOfRole = array_map(fn($v) => $v['grantee_name'], $ownershipOfRole);
-
-            $grantsOfRoles = array_merge($grantsOfRoles, array_unique($ownershipOfRole));
-        }
-
-        $uniqueMainRoles = array_unique($grantsOfRoles);
-
-        assert(count($uniqueMainRoles) === 1);
-
-        $mainRole = current($uniqueMainRoles);
-
-        return [
-            'name' => $mainRole,
-            'assignedGrants' => $this->sourceConnection->fetchAll(sprintf(
-                'SHOW GRANTS TO ROLE %s;',
-                Helper::quoteIdentifier($mainRole)
-            )),
-        ];
-    }
-
-    public function exportRolesGrants(): array
-    {
-        $tmp = [];
-        foreach ($this->databases as $database) {
-            $databaseRole = $this->sourceConnection->getOwnershipRoleOnDatabase($database);
-
-            $roles = $this->sourceConnection->fetchAll(sprintf(
-                'SHOW ROLES LIKE %s',
-                QueryBuilder::quote($databaseRole)
-            ));
-
-            $roles = $this->getOtherRolesToMainProjectRole($roles);
-
-            foreach ($roles as $role) {
-                $tmp[$databaseRole][] = array_merge(
-                    $role,
-                    [
-                        'assignedGrants' => $this->sourceConnection->fetchAll(sprintf(
-                            'SHOW GRANTS TO ROLE %s;',
-                            Helper::quoteIdentifier($role['name'])
-                        )),
-                        'assignedFutureGrants' => $this->sourceConnection->fetchAll(sprintf(
-                            'SHOW FUTURE GRANTS TO ROLE %s;',
-                            Helper::quoteIdentifier($role['name'])
-                        )),
-                    ]
-                );
-            }
-        }
-        return $tmp;
-    }
-
-    public function createMainRole(array $mainRoleWithGrants, array $userPasswords): void
-    {
-        $user = $mainRole = $mainRoleWithGrants['name'];
-
-        $this->destinationConnection->createRole(
-            ['name' => $mainRole, 'privilege' => 'OWNERSHIP'],
-            $this->config->getTargetSnowflakeUser()
-        );
-
-        $mainRoleGrants = [
-            'GRANT CREATE DATABASE ON ACCOUNT TO ROLE %s;',
-            'GRANT CREATE ROLE ON ACCOUNT TO ROLE %s WITH GRANT OPTION;',
-            'GRANT CREATE USER ON ACCOUNT TO ROLE %s WITH GRANT OPTION;',
-        ];
-
-        foreach ($mainRoleGrants as $mainRoleGrant) {
-            $this->destinationConnection->query(sprintf($mainRoleGrant, $mainRole));
-        }
-
-        $warehouses = array_filter($mainRoleWithGrants['assignedGrants'], fn($v) => $v['granted_on'] === 'WAREHOUSE');
-        foreach ($warehouses as $warehouse) {
-            $this->createWarehouse($warehouse);
-            $this->destinationConnection->assignGrantToRole($warehouse);
-        }
-
-        if (isset($userPasswords[$user])) {
-            $password = $userPasswords[$user];
-        } else {
-            $password = Helper::generateRandomString();
-        }
-
-        $this->destinationConnection->query(sprintf(
-            'CREATE USER IF NOT EXISTS %s PASSWORD=%s DEFAULT_ROLE=%s',
-            $user,
-            QueryBuilder::quote($password),
-            $mainRole
-        ));
-
-        $this->destinationConnection->query(sprintf(
-            'GRANT ROLE %s TO USER %s;',
-            $mainRole,
-            $user
-        ));
-
-        $projectUsers = array_filter($mainRoleWithGrants['assignedGrants'], function ($v) {
-            if ($v['privilege'] !== 'OWNERSHIP') {
-                return false;
-            }
-            if ($v['granted_on'] !== 'USER') {
-                return false;
-            }
-            return in_array($v['name'], $this->databases);
-        });
-
-        $sourceGrants = $this->sourceConnection->fetchAll(sprintf(
-            'SHOW GRANTS TO ROLE %s',
-            Helper::quoteIdentifier($this->sourceConnection->getCurrentRole())
-        ));
-
-        $sourceGrants = array_filter(
-            $sourceGrants,
-            fn($v) => $v['granted_on'] === 'WAREHOUSE' && $v['privilege'] === 'USAGE'
-        );
-        assert(count($sourceGrants) > 0);
-
-        foreach ($projectUsers as $projectUser) {
-            $this->createUser($projectUser, $userPasswords);
-            $this->destinationConnection->assignGrantToRole($projectUser);
-        }
-
-        $this->destinationConnection->useRole($this->mainMigrationRoleTargetAccount);
-
-        $this->destinationConnection->query(sprintf(
-            'GRANT ROLE %s TO ROLE SYSADMIN;',
-            $mainRole
-        ));
-    }
-
-    private function createWarehouse(array $warehouse): string
-    {
-        $role = $this->sourceConnection->getCurrentRole();
-        $this->destinationConnection->useRole($this->mainMigrationRoleTargetAccount);
-        $this->sourceConnection->useRole($this->mainMigrationRoleSourceAccount);
-        $warehouseInfo = $this->sourceConnection->fetchAll(sprintf(
-            'SHOW WAREHOUSES LIKE %s',
-            QueryBuilder::quote($warehouse['name'])
-        ));
-        assert(count($warehouseInfo) === 1);
-
-        $warehouseInfo = current($warehouseInfo);
-
-        $sqlTemplate = <<<SQL
-CREATE WAREHOUSE IF NOT EXISTS %s
-    WITH WAREHOUSE_SIZE = %s
-        WAREHOUSE_TYPE = %s
-        AUTO_SUSPEND = %s
-        AUTO_RESUME = %s
-;
-SQL;
-
-        $sql = sprintf(
-            $sqlTemplate,
-            Helper::quoteIdentifier($warehouseInfo['name']),
-            QueryBuilder::quote($warehouseInfo['size']),
-            QueryBuilder::quote($warehouseInfo['type']),
-            $warehouseInfo['auto_suspend'],
-            $warehouseInfo['auto_resume']
-        );
-
-        $this->destinationConnection->query($sql);
-
-        $this->sourceConnection->useRole($role);
-        return $warehouseInfo['size'];
-    }
-
-    public function grantRoleToUsers(): void
-    {
-        $this->destinationConnection->useRole($this->mainMigrationRoleTargetAccount);
-        $this->sourceConnection->useRole($this->mainMigrationRoleSourceAccount);
-
-        foreach ($this->usedUsers as $user) {
-            $grants = $this->sourceConnection->fetchAll(sprintf(
-                'SHOW GRANTS TO USER %s',
-                Helper::quoteIdentifier($user)
-            ));
-
-            foreach ($grants as $grant) {
-                $this->destinationConnection->useRole($grant['granted_by']);
-                $this->destinationConnection->query(sprintf(
-                    'GRANT ROLE %s TO %s %s',
-                    $grant['role'],
-                    $grant['granted_to'],
-                    $grant['grantee_name'],
-                ));
-            }
-        }
-    }
-
-    public function postMigrationCleanup(): void
-    {
-        foreach ($this->databases as $database) {
-            $databases = [
-                $database . '_OLD',
-                $database . '_SHARE',
-            ];
-
-            $removeDatabases = array_filter(
-                $this->destinationConnection->fetchAll('SHOW DATABASES;'),
-                fn($v) => in_array($v['name'], $databases)
-            );
-
-            foreach ($removeDatabases as $removeDatabase) {
-                $this->destinationConnection->useRole($removeDatabase['owner']);
-                $this->destinationConnection->query(sprintf(
-                    'DROP DATABASE IF EXISTS %s',
-                    Helper::quoteIdentifier($removeDatabase['name'])
-                ));
-            }
-
-            $userRoles = $this->destinationConnection->fetchAll(sprintf(
-                'SHOW GRANTS TO USER %s',
-                Helper::quoteIdentifier($this->config->getTargetSnowflakeUser())
-            ));
-
-            foreach (array_reverse($userRoles) as $userRole) {
-                if ($userRole['role'] === $this->mainMigrationRoleTargetAccount) {
-                    continue;
-                }
-                try {
-                    if ($userRole['granted_by']) {
-                        $this->destinationConnection->useRole($userRole['granted_by']);
-                    } else {
-                        $this->destinationConnection->useRole($this->mainMigrationRoleTargetAccount);
-                    }
-                    $this->destinationConnection->query(sprintf(
-                        'REVOKE ROLE %s FROM USER %s',
-                        Helper::quoteIdentifier($userRole['role']),
-                        Helper::quoteIdentifier($userRole['grantee_name']),
-                    ));
-                } catch (RuntimeException $e) {
-                    $this->logger->info(sprintf(
-                        'Warning: Query failed, please check manually: %s',
-                        $e->getMessage()
-                    ));
-                }
-            }
-        }
-    }
-
-    public function postMigrationCheckStructure(array $mainRoleWithGrants): void
-    {
-        $warehouses = array_filter($mainRoleWithGrants['assignedGrants'], fn($v) => $v['granted_on'] === 'WAREHOUSE');
-
-        $useWarehouse = sprintf(
-            'USE WAREHOUSE %s',
-            Helper::quoteIdentifier(current($warehouses)['name'])
-        );
-        $this->sourceConnection->query($useWarehouse);
-        $this->destinationConnection->query($useWarehouse);
-
-        foreach ($this->databases as $database) {
-            $this->logger->info(sprintf('Checking database %s', $database));
-            $databaseRole = $this->sourceConnection->getOwnershipRoleOnDatabase($database);
-
-            $this->logger->info(sprintf('Getting roles and users for database %s', $database));
-            $rolesAndUsers = $this->listRolesAndUsers($databaseRole);
-            $rolesAndUsers = array_merge_recursive(
-                $rolesAndUsers,
-                ['users' => [$databaseRole], 'roles' => [$databaseRole]]
-            );
-
-            $compares = [];
-            // phpcs:disable Generic.Files.LineLength
-//            Compare TABLES
-            $compares[] = [
-                'group' => 'Tables',
-                'itemNameKey' => 'TABLE_NAME',
-                'sql' => sprintf(
-                    'SELECT %s FROM SNOWFLAKE.ACCOUNT_USAGE.TABLES WHERE DELETED IS NULL AND TABLE_CATALOG = %s ORDER BY TABLE_SCHEMA, TABLE_NAME;',
-                    implode(',', [
-                        'CONCAT(TABLE_SCHEMA, \'.\', TABLE_NAME) AS ID',
-                        'TABLE_NAME',
-                        'TABLE_SCHEMA',
-                        'TABLE_OWNER',
-                        'TABLE_TYPE',
-                        'ROW_COUNT',
-                        // 'BYTES',
-                    ]),
-                    QueryBuilder::quote($database)
-                ),
-            ];
-
-//            Compare USERS
-            $compares[] = [
-                'group' => 'Users',
-                'itemNameKey' => 'NAME',
-                'sql' => sprintf(
-                    'SELECT %s FROM SNOWFLAKE.ACCOUNT_USAGE.USERS WHERE DELETED_ON IS NULL AND NAME IN (%s) ORDER BY NAME;',
-                    implode(',', [
-                        'NAME AS ID',
-                        'NAME',
-                        'LOGIN_NAME',
-                        'DISPLAY_NAME',
-                        'FIRST_NAME',
-                        'LAST_NAME',
-                        'EMAIL',
-                        'DEFAULT_WAREHOUSE',
-                        'DEFAULT_NAMESPACE',
-                        'DEFAULT_ROLE',
-                        'OWNER',
-                    ]),
-                    implode(', ', array_map(fn($v) => QueryBuilder::quote($v), $rolesAndUsers['roles']))
-                ),
-            ];
-
-//            Compare GRANTS TO USERS
-            $compares[] = [
-                'group' => 'Grants to user',
-                'itemNameKey' => 'ROLE',
-                'sql' => sprintf(
-                    'SELECT %s FROM SNOWFLAKE.ACCOUNT_USAGE.GRANTS_TO_USERS WHERE DELETED_ON IS NULL AND ROLE IN (%s) ORDER BY ROLE, GRANTED_BY;',
-                    implode(',', [
-                        'CONCAT(ROLE, GRANTED_TO, GRANTEE_NAME, GRANTED_BY) AS ID',
-                        'ROLE',
-                        'GRANTED_TO',
-                        'GRANTEE_NAME',
-                        'GRANTED_BY',
-                    ]),
-                    implode(', ', array_map(fn($v) => QueryBuilder::quote($v), $rolesAndUsers['users']))
-                ),
-            ];
-
-//            Compare ROLES
-            $compares[] = [
-                'group' => 'Roles',
-                'itemNameKey' => 'NAME',
-                'sql' => sprintf(
-                    'SELECT %s FROM SNOWFLAKE.ACCOUNT_USAGE.ROLES WHERE DELETED_ON IS NULL AND NAME IN (%s) ORDER BY NAME;',
-                    implode(',', [
-                        'CONCAT(NAME, OWNER) AS ID',
-                        'NAME',
-                        'COMMENT',
-                        'OWNER',
-                    ]),
-                    implode(', ', array_map(fn($v) => QueryBuilder::quote($v), $rolesAndUsers['roles']))
-                ),
-            ];
-
-//            Compare GRANTS TO ROLES
-            $compares[] = [
-                'group' => 'Grants to roles',
-                'itemNameKey' => 'ID',
-                'sql' => sprintf(
-                    'SELECT %s FROM SNOWFLAKE.ACCOUNT_USAGE.GRANTS_TO_ROLES WHERE DELETED_ON IS NULL AND GRANTEE_NAME IN (%s) ORDER BY GRANTEE_NAME, PRIVILEGE, NAME;',
-                    implode(',', [
-                        'CONCAT(PRIVILEGE, GRANTED_ON, NAME, GRANTED_TO, GRANTEE_NAME, GRANTED_BY, CASE WHEN TABLE_CATALOG IS NULL THEN \'\' ELSE TABLE_CATALOG END, CASE WHEN TABLE_SCHEMA IS NULL THEN \'\' ELSE TABLE_SCHEMA END) AS ID',
-                        'PRIVILEGE',
-                        'GRANTED_ON',
-                        'NAME',
-                        'TABLE_CATALOG',
-                        'TABLE_SCHEMA',
-                        'GRANTED_TO',
-                        'GRANTEE_NAME',
-                        'GRANT_OPTION',
-                        'GRANTED_BY',
-                    ]),
-                    implode(', ', array_map(fn($v) => QueryBuilder::quote($v), $rolesAndUsers['roles']))
-                ),
-            ];
-            // phpcs:enable Generic.Files.LineLength
-
-            foreach ($compares as $compare) {
-                $this->compareData($compare['group'], $compare['itemNameKey'], $compare['sql']);
-            }
-        }
-    }
-
-    public function postMigrationCheckData(array $mainRoleWithGrants): void
-    {
-        $sourceRegion = $this->sourceConnection->fetchAll(sprintf(
-            'SHOW regions LIKE %s',
-            QueryBuilder::quote($this->sourceConnection->getRegion())
-        ));
-
-        $targetRegion = $this->destinationConnection->fetchAll(sprintf(
-            'SHOW regions LIKE %s',
-            QueryBuilder::quote($this->destinationConnection->getRegion())
-        ));
-
-        $sourceAccount = sprintf(
-            '%s.%s',
-            $this->sourceConnection->getAccount(),
-            $sourceRegion[0]['region']
-        );
-        $targetAccount = sprintf(
-            '%s.%s',
-            $this->destinationConnection->getAccount(),
-            $targetRegion[0]['region']
-        );
-
-        $warehousesGrant = array_values(array_filter(
-            $mainRoleWithGrants['assignedGrants'],
-            fn($v) => $v['granted_on'] === 'WAREHOUSE' && $v['privilege'] === 'USAGE'
-        ));
-        assert(count($warehousesGrant) > 0);
-        $this->sourceConnection->grantRoleToUser($this->config->getSourceSnowflakeUser(), $mainRoleWithGrants['name']);
-        $this->destinationConnection->grantRoleToUser(
-            $this->config->getTargetSnowflakeUser(),
-            $mainRoleWithGrants['name']
-        );
-
-        foreach ($this->config->getDatabases() as $database) {
-            $schemas = $this->sourceConnection->fetchAll(sprintf(
-                'SHOW SCHEMAS IN DATABASE %s',
-                Helper::quoteIdentifier($database)
-            ));
-            foreach ($schemas as $schema) {
-                if (in_array($schema['name'], ['INFORMATION_SCHEMA', 'PUBLIC'])) {
-                    continue;
-                }
-                $this->sourceConnection->useRole($mainRoleWithGrants['name']);
-                $this->destinationConnection->useRole($mainRoleWithGrants['name']);
-                $this->sourceConnection->grantRoleToUser($this->config->getSourceSnowflakeUser(), $schema['owner']);
-                $this->destinationConnection->grantRoleToUser(
-                    $this->config->getTargetSnowflakeUser(),
-                    $schema['owner']
-                );
-                $this->sourceConnection->useRole($schema['owner']);
-
-                $tables = $this->sourceConnection->fetchAll(sprintf(
-                    'SHOW TABLES IN SCHEMA %s.%s',
-                    Helper::quoteIdentifier($database),
-                    Helper::quoteIdentifier($schema['name'])
-                ));
-                foreach ($tables as $table) {
-                    $primaryKeys = $this->sourceConnection->fetchAll(sprintf(
-                        'SHOW PRIMARY KEYS IN TABLE %s.%s.%s',
-                        Helper::quoteIdentifier($database),
-                        Helper::quoteIdentifier($schema['name']),
-                        Helper::quoteIdentifier($table['name'])
-                    ));
-                    if (!$primaryKeys) {
-                        $this->logger->info(sprintf(
-                            'Warning: Table %s.%s.%s has no primary key. Skipping',
-                            $database,
-                            $schema['name'],
-                            $table['name']
-                        ));
-                        continue;
-                    }
-                    $primaryKeys = array_map(fn($v) => $v['column_name'], $primaryKeys);
-
-                    $columns = $this->sourceConnection->fetchAll(sprintf(
-                        'SHOW COLUMNS IN TABLE %s.%s.%s',
-                        Helper::quoteIdentifier($database),
-                        Helper::quoteIdentifier($schema['name']),
-                        Helper::quoteIdentifier($table['name'])
-                    ));
-                    $columns = array_map(fn($v) => $v['column_name'], $columns);
-                    $columns = array_filter($columns, fn($v) => !in_array($v, $primaryKeys));
-                    $columns = array_filter($columns, fn($v) => $v !== '_timestamp');
-
-                    $arguments = [
-                        'sourceAccount' => $sourceAccount,
-                        'sourceUser' => $this->config->getSourceSnowflakeUser(),
-                        'sourcePassword' => $this->config->getSourceSnowflakePassword(),
-                        'targetAccount' => $targetAccount,
-                        'targetUser' => $this->config->getTargetSnowflakeUser(),
-                        'targetPassword' => $this->config->getTargetSnowflakePassword(),
-                        'role' => $schema['owner'],
-                        'warehouse' => $warehousesGrant[0]['name'],
-                        'database' => $database,
-                        'schema' => $schema['name'],
-                        'table' => $table['name'],
-                        'extraColumns' => implode(',', $columns),
-                        'primaryKeys' => implode(',', $primaryKeys),
-                    ];
-
-                    array_walk($arguments, function (&$value, $key): void {
-                        $value = sprintf('--%s "%s"', $key, $value);
-                    });
-
-                    $process = Process::fromShellCommandline(
-                        'python3 ' . __DIR__ . '/dataDiff.py ' . implode(' ', $arguments)
-                    );
-                    $process->setTimeout(null);
-                    $process->run();
-                    if (!$process->isSuccessful()) {
-                        $this->logger->info(sprintf(
-                            'Warning: Checking table "%s.%s.%s" ends with error: "%s"',
-                            $database,
-                            $schema['name'],
-                            $table['name'],
-                            $process->getOutput()
-                        ));
-                    } else {
-                        $this->logger->info(sprintf(
-                            'Checking table "%s.%s.%s" ends successfully. %s',
-                            $database,
-                            $schema['name'],
-                            $table['name'],
-                            $process->getOutput()
-                        ));
-                    }
-                }
-            }
-        }
-    }
-
-    public function printUnusedGrants(array $grants): void
-    {
-        foreach ($grants as $databaseGrants) {
-            [
-                'grants' => [
-                    'other' => $otherGrants,
-                ],
-                'futureGrants' => [
-                    'other' => $otherFutureGrants,
-                ],
-            ] = Helper::parseGrantsToObjects($databaseGrants);
-
-            foreach ($otherGrants as $grant) {
-                $this->logger->alert(sprintf(
-                    'Unused grant "%s": GRANT %s ON %s %s TO %s %s %s',
-                    $grant['name'],
-                    $grant['privilege'],
-                    $grant['granted_on'],
-                    $grant['granted_on'] !== 'ACCOUNT' ? $grant['name'] : '',
-                    $grant['granted_to'],
-                    Helper::quoteIdentifier($grant['grantee_name']),
-                    $grant['grant_option'] === 'true' ? 'WITH GRANT OPTION' : '',
-                ));
-            }
-
-            foreach ($otherFutureGrants as $grant) {
-                $this->logger->alert(sprintf(
-                    'Unused FUTURE grant "%s": GRANT %s ON FUTURE TABLES IN SCHEMA %s TO ROLE %s %s',
-                    $grant['name'],
-                    $grant['privilege'],
-                    $grant['name'],
-                    Helper::quoteIdentifier($grant['grantee_name']),
-                    $grant['grant_option'] === 'true' ? 'WITH GRANT OPTION' : '',
-                ));
-            }
-        }
-    }
-
-    private function compareData(string $group, string $itemNameKey, string $sql): void
-    {
-        $this->logger->info(sprintf('Getting source data for "%s".', $group));
-        $sourceData = $this->sourceConnection->fetchAll($sql);
-        $this->logger->info(sprintf('Getting target data for "%s".', $group));
-        $targetData = $this->destinationConnection->fetchAll($sql);
-
-        if (count($sourceData) !== count($targetData)) {
-            $this->logger->info(sprintf(
-                'Alert - %s: Source data count (%s) does not equal target data count (%s)',
-                $group,
-                count($sourceData),
-                count($targetData)
-            ));
-        }
-        $sourceData = array_combine(array_map(fn($v) => $v['ID'], $sourceData), $sourceData);
-        $targetData = array_combine(array_map(fn($v) => $v['ID'], $targetData), $targetData);
-
-        array_walk($sourceData, fn(&$v) => $v = serialize($v));
-        array_walk($targetData, fn(&$v) => $v = serialize($v));
-
-        $diffs = [
-            array_diff($sourceData, $targetData),
-            array_diff($targetData, $sourceData),
-        ];
-        $print = [];
-        foreach ($diffs as $diff) {
-            foreach ($diff as $k => $serializeItem) {
-                if (!isset($sourceData[$k])) {
-                    $this->logger->info(sprintf(
-                        'Alert - %s: Item "%s" doesn\'t exists in source account',
-                        $group,
-                        $k
-                    ));
-                    continue;
-                }
-                if (!isset($targetData[$k])) {
-                    $this->logger->info(sprintf(
-                        'Alert - %s: Item "%s" doesn\'t exists in target account',
-                        $group,
-                        $k
-                    ));
-                    continue;
-                }
-                $itemSource = (array) unserialize($sourceData[$k]);
-                $itemName = strval($itemSource[$itemNameKey]);
-                if (in_array($itemName, $print)) {
-                    continue;
-                }
-                $itemTarget = (array) unserialize($targetData[$k]);
-                $itemDiffs = [
-                    'target' => array_diff($itemSource, $itemTarget),
-                    'source' => array_diff($itemTarget, $itemSource),
-                ];
-
-                foreach ($itemDiffs as $missingIn => $itemDiff) {
-                    $this->printDiffAlert($group, $itemName, $missingIn, $itemDiff);
-                }
-
-                $print[] = $itemName;
-            }
-        }
-    }
-
-    private function printDiffAlert(string $group, string $name, string $missingIn, array $data): void
-    {
-        if (!$data) {
-            return;
-        }
-        array_walk($data, fn(&$v, $k) => $v = sprintf('%s: %s', $k, $v));
-
-        $this->logger->info(sprintf(
-            'Alert - %s: "%s" is not same. Missing in %s account (%s)',
-            $group,
-            $name,
-            $missingIn,
-            implode(';', $data)
-        ));
-    }
-
-    private function listRolesAndUsers(string $role): array
-    {
-        $grants = $this->destinationConnection->fetchAll(sprintf(
-            'SHOW GRANTS TO ROLE %s',
-            Helper::quoteIdentifier($role)
-        ));
-
-        $filteredGrants = array_filter(
-            $grants,
-            fn($v) => $v['privilege'] === 'OWNERSHIP' && (in_array($v['granted_on'], ['USER', 'ROLE']))
-        );
-
-        $tmp = [
-            'users' => [],
-            'roles' => [],
-        ];
-        foreach ($filteredGrants as $filteredGrant) {
-            switch ($filteredGrant['granted_on']) {
-                case 'USER':
-                    $tmp['users'][] = $filteredGrant['name'];
-                    break;
-                case 'ROLE':
-                    $tmp['roles'][] = $filteredGrant['name'];
-                    $childRoles = $this->listRolesAndUsers($filteredGrant['name']);
-                    $tmp = array_merge_recursive($tmp, $childRoles);
-                    break;
-            }
-        }
-
-        return $tmp;
-    }
-
-    private function createUser(array $userGrant, array $passwordOfUsers): void
-    {
-        $this->destinationConnection->useRole($userGrant['granted_by']);
-
-        if (!in_array($userGrant['name'], $this->usedUsers)) {
-            $this->usedUsers[] = $userGrant['name'];
-        }
-
-        $this->sourceConnection->useRole($this->mainMigrationRoleSourceAccount);
-        $describeUser = $this->sourceConnection->fetchAll(sprintf(
-            'SHOW USERS LIKE %s',
-            QueryBuilder::quote($userGrant['name'])
-        ));
-        assert(count($describeUser) === 1);
-
-        $allowOptions = [
-            'default_secondary_roles',
-            'default_role',
-            'default_namespace',
-            'default_warehouse',
-            'display_name',
-            'login_name',
-        ];
-
-        $describeUser = (array) array_filter(
-            current($describeUser),
-            fn($k) => in_array($k, $allowOptions),
-            ARRAY_FILTER_USE_KEY
-        );
-
-        $describeUser = array_filter(
-            $describeUser,
-            fn($v) => $v !== ''
-        );
-
-        array_walk(
-            $describeUser,
-            fn(&$item, $k) => $item = sprintf('%s = %s', strtoupper($k), QueryBuilder::quote($item))
-        );
-
-        if (isset($passwordOfUsers[$userGrant['name']])) {
-            $password = $passwordOfUsers[$userGrant['name']];
-        } else {
-            $password = Helper::generateRandomString();
-        }
-
-        $describeUser['password'] = sprintf(
-            'PASSWORD = %s',
-            QueryBuilder::quote($password)
-        );
-
-        $this->destinationConnection->query(sprintf(
-            'CREATE USER %s %s',
-            $userGrant['name'],
-            implode(' ', $describeUser),
-        ));
-    }
-
-    private function getOtherRolesToMainProjectRole(array $roles): array
-    {
-        foreach ($roles as $role) {
-            $grantsToRole = $this->sourceConnection->fetchAll(sprintf(
-                'SHOW GRANTS TO ROLE %s;',
-                Helper::quoteIdentifier($role['name'])
-            ));
-
-            $ownershipToRole = array_filter($grantsToRole, fn($v) => $v['privilege'] === 'OWNERSHIP');
-
-            $rolesInRole = array_filter($ownershipToRole, fn($v) => $v['granted_on'] === 'ROLE');
-            $filteredRolesInRole = array_filter($rolesInRole, fn($v) => !in_array($v['name'], $roles));
-            $roles = array_merge($roles, array_combine(
-                array_map(fn($v) => $v['name'], $filteredRolesInRole),
-                $filteredRolesInRole
-            ));
-        }
-
-        return $roles;
-    }
-
-    private function assignSharePrivilegesToRole(string $database, string $role): void
-    {
-        $this->destinationConnection->assignGrantToRole([
-            'privilege' => 'IMPORTED PRIVILEGES',
-            'granted_on' => 'DATABASE',
-            'name' => Helper::quoteIdentifier($database . '_SHARE'),
-            'granted_to' => 'ROLE',
-            'grantee_name' => $role,
-            'grant_option' => 'false',
-            'granted_by' => $this->mainMigrationRoleTargetAccount,
-        ]);
-    }
-
-    private function getDataToRemove(string $role): array
-    {
-        $grants = $this->destinationConnection->fetchAll(sprintf(
-            'SHOW GRANTS TO ROLE %s',
-            Helper::quoteIdentifier($role)
-        ));
-
-        $filteredGrants = array_filter($grants, function ($v) {
-            $usageWarehouse = $v['privilege'] === 'USAGE' && $v['granted_on'] === 'WAREHOUSE';
-            $ownership = $v['privilege'] === 'OWNERSHIP' && (in_array($v['granted_on'], ['USER', 'ROLE']));
-
-            return $ownership || $usageWarehouse;
-        });
-
-        $mapGrants = [];
-        foreach ($filteredGrants as $filteredGrant) {
-            $mapGrants[$filteredGrant['granted_on']][$filteredGrant['name']] = $filteredGrant;
-        }
-
-        if (isset($mapGrants['ROLE'])) {
-            $roleGrants = $mapGrants['ROLE'];
-            foreach ($roleGrants as $roleGrant) {
-                $mapGrants = array_merge_recursive(
-                    $this->getDataToRemove($roleGrant['name']),
-                    $mapGrants,
-                );
-            }
-        }
-
-        return $mapGrants;
-    }
-
-    private function canCloneTable(string $mainRole, string $database, string $schema, string $table): bool
-    {
-
-        $sqlTemplate = 'SELECT max("_timestamp") as "maxTimestamp" FROM %s.%s.%s';
-
-        $currentRole = $this->destinationConnection->getCurrentRole();
-        try {
-            $this->destinationConnection->useRole($mainRole);
-
-            $lastUpdateTableInOldDatabase = $this->destinationConnection->fetchAll(sprintf(
-                $sqlTemplate,
-                Helper::quoteIdentifier($database . '_OLD'),
-                Helper::quoteIdentifier($schema),
-                Helper::quoteIdentifier($table)
-            ));
-
-            $this->destinationConnection->useRole($this->mainMigrationRoleTargetAccount);
-            $lastUpdateTableInShareDatabase = $this->destinationConnection->fetchAll(sprintf(
-                $sqlTemplate,
-                Helper::quoteIdentifier($database . '_SHARE'),
-                Helper::quoteIdentifier($schema),
-                Helper::quoteIdentifier($table)
-            ));
-        } catch (RuntimeException $e) {
-            return false;
-        } finally {
-            $this->destinationConnection->useRole($currentRole);
-        }
-
-        return $lastUpdateTableInOldDatabase[0]['maxTimestamp'] === $lastUpdateTableInShareDatabase[0]['maxTimestamp'];
-    }
-
-    private function grantUsageToOldTable(
-        string $database,
-        string $schema,
-        string $table,
-        string $mainRole,
-        string $role
-    ): void {
-        $currentRole = $this->destinationConnection->getCurrentRole();
-        $this->destinationConnection->useRole($mainRole);
-        $dbExists = $this->destinationConnection->fetchAll(sprintf(
-            'SHOW DATABASES LIKE %s',
-            QueryBuilder::quote($database . '_OLD')
-        ));
-        if (!$dbExists) {
-            return;
-        }
-
-        $sqls = [
-            sprintf(
-                'grant usage on database %s to role %s;',
-                Helper::quoteIdentifier($database . '_OLD'),
-                Helper::quoteIdentifier($role)
-            ),
-            sprintf(
-                'grant usage on schema %s.%s to role %s;',
-                Helper::quoteIdentifier($database . '_OLD'),
-                Helper::quoteIdentifier($schema),
-                Helper::quoteIdentifier($role)
-            ),
-            sprintf(
-                'grant select on table %s.%s.%s to role %s;',
-                Helper::quoteIdentifier($database . '_OLD'),
-                Helper::quoteIdentifier($schema),
-                Helper::quoteIdentifier($table),
-                Helper::quoteIdentifier($role)
-            ),
-        ];
-
-        foreach ($sqls as $sql) {
-            $this->destinationConnection->query($sql);
-        }
-        $this->destinationConnection->useRole($currentRole);
-    }
-
-    private function buildSqlFunctionQuery(array $function, array $functionParams): string
-    {
-        $sql = <<<SQL
-CREATE %s FUNCTION %s%s
-returns %s 
-AS 
-$$
-%s
-$$
-;
-SQL;
-
-        return sprintf(
-            $sql,
-            $function['is_secure'] === 'Y' ? 'SECURE' : '',
-            Helper::quoteIdentifier($function['name']),
-            $functionParams['signature'],
-            $functionParams['returns'],
-            trim($functionParams['body'])
-        );
-    }
-
-    private function buildPythonFunctionQuery(array $function, array $functionParams): string
-    {
-        $python = <<<SQL
-CREATE FUNCTION %s%s
-RETURNS %s 
-LANGUAGE PYTHON
-RUNTIME_VERSION=%s
-HANDLER=%s
-%s
-%s
-AS 
-$$
-%s
-$$
-;
-SQL;
-
-        return sprintf(
-            $python,
-            Helper::quoteIdentifier($function['name']),
-            $functionParams['signature'],
-            $functionParams['returns'],
-            $functionParams['runtime_version'],
-            QueryBuilder::quote($functionParams['handler']),
-            $functionParams['imports'] !== '[]' ?
-                'IMPORTS = (' . substr($functionParams['imports'], 1, -1) . ')' :
-                '',
-            $functionParams['packages'] !== '[]' ?
-                'PACKAGES = (' . substr($functionParams['packages'], 1, -1) . ')' :
-                '',
-            trim($functionParams['body'])
-        );
-    }
-
-    private function assignForeignGrants(array $grants, string $databaseRole): void
-    {
-        $foreignGrants = array_filter($grants, fn($v) => $v['granted_by'] !== $databaseRole);
-
-        array_walk($foreignGrants, fn($grant) => $this->destinationConnection->assignGrantToRole($grant));
-    }
-
-    private function copyFunctions(string $database, array $functionsGrants): void
-    {
-        $this->logger->info(sprintf('Cloning functions from database "%s"', $database));
-        $functions = $this->sourceConnection->fetchAll(sprintf(
-            'SHOW FUNCTIONS IN DATABASE %s;',
-            Helper::quoteIdentifier($database),
-        ));
-
-        $functions = array_filter($functions, fn($v) => $v['catalog_name'] === $database);
-
-        $this->destinationConnection->query('USE DATABASE ' . Helper::quoteIdentifier($database) . ';');
-        foreach ($functions as $function) {
-            preg_match('/.*\((.*)\) RETURN/', $function['arguments'], $matches);
-            $descFunction = $this->sourceConnection->fetchAll(sprintf(
-                'DESC FUNCTION %s.%s.%s(%s)',
-                Helper::quoteIdentifier($function['catalog_name']),
-                Helper::quoteIdentifier($function['schema_name']),
-                Helper::quoteIdentifier($function['name']),
-                $matches[1]
-            ));
-
-            $functionParams = array_combine(
-                array_map(fn($v) => $v['property'], $descFunction),
-                array_map(fn($v) => $v['value'], $descFunction)
-            );
-
-            switch ($function['language']) {
-                case 'SQL':
-                    $functionQuery = $this->buildSqlFunctionQuery($function, $functionParams);
-                    break;
-                case 'PYTHON':
-                    $functionQuery = $this->buildPythonFunctionQuery($function, $functionParams);
-                    break;
-                default:
-                    $this->logger->warning(sprintf(
-                        'Warning: Skip creating function "%s". Language "%s" is not supported.',
-                        Helper::quoteIdentifier($function['name']),
-                        $function['language']
-                    ));
-                    continue 2;
-            }
-
-            $ownership = array_filter(
-                $functionsGrants,
-                fn($v) => str_contains($v['granted_by'], $function['schema_name'])
-            );
-
-            $this->destinationConnection->useRole(current($ownership)['granted_by']);
-
-            $this->destinationConnection->query(sprintf(
-                'USE SCHEMA %s.%s;',
-                Helper::quoteIdentifier($function['catalog_name']),
-                Helper::quoteIdentifier($function['schema_name'])
-            ));
-
-            try {
-                $this->destinationConnection->query($functionQuery);
-            } catch (Throwable $e) {
-                $this->logger->info(sprintf(
-                    'Skip creating function %s.%s.%s. Error: "%s".',
-                    Helper::quoteIdentifier($function['catalog_name']),
-                    Helper::quoteIdentifier($function['schema_name']),
-                    Helper::quoteIdentifier($function['name']),
-                    $e->getMessage()
-                ));
-            }
-        }
-
-        foreach ($functionsGrants as $functionsGrant) {
-            if ($functionsGrant['privilege'] === 'OWNERSHIP') {
-                continue;
-            }
-            $this->destinationConnection->assignGrantToRole($functionsGrant);
-        }
-    }
-
-    private function copyViews(string $database, string $databaseRole, array $viewsGrants): void
-    {
-        $this->logger->info(sprintf('Cloning views from database "%s"', $database));
-        $this->sourceConnection->grantRoleToUser($this->config->getSourceSnowflakeUser(), $databaseRole);
-        $this->sourceConnection->useRole($databaseRole);
-        $views = $this->sourceConnection->fetchAll(sprintf(
-            'SHOW VIEWS IN DATABASE %s;',
-            Helper::quoteIdentifier($database),
-        ));
-
-        $views = array_filter($views, fn($v) => !in_array($v['schema_name'], self::SKIP_CLONE_SCHEMAS));
-
-        $try = 0;
-        while ($views && $try < 5) {
-            if ($try > 0) {
-                $this->logger->info(sprintf(
-                    'Attempting to create %s failed views.',
-                    count($views)
-                ));
-            }
-            foreach ($views as $viewKey => $view) {
-                $this->destinationConnection->useRole($view['owner']);
-
-                $this->destinationConnection->query(sprintf(
-                    'USE SCHEMA %s.%s;',
-                    Helper::quoteIdentifier($view['database_name']),
-                    Helper::quoteIdentifier($view['schema_name'])
-                ));
-
-                try {
-                    $this->destinationConnection->query($view['text']);
-                    unset($views[$viewKey]);
-                } catch (Throwable $e) {
-                    $this->logger->info(sprintf(
-                        'Warning: Skip creating view %s.%s.%s. Error: "%s".',
-                        Helper::quoteIdentifier($view['database_name']),
-                        Helper::quoteIdentifier($view['schema_name']),
-                        Helper::quoteIdentifier($view['name']),
-                        $e->getMessage()
-                    ));
-                }
-            }
-            $try++;
-        }
-
-        foreach ($viewsGrants as $viewsGrant) {
-            if ($viewsGrant['privilege'] === 'OWNERSHIP') {
-                continue;
-            }
-            $this->destinationConnection->assignGrantToRole($viewsGrant);
-        }
-    }
-
-    private function copyProcedures(string $database, array $proceduresGrants): void
-    {
-        $this->logger->info(sprintf('Cloning procedures from database "%s"', $database));
-
-        $procedures = $this->sourceConnection->fetchAll(sprintf(
-            'SHOW PROCEDURES IN DATABASE %s;',
-            Helper::quoteIdentifier($database),
-        ));
-
-        $procedures = array_filter($procedures, fn($v) => $v['catalog_name'] === $database);
-
-        foreach ($procedures as $procedure) {
-            preg_match('/.*\((.*)\) RETURN/', $procedure['arguments'], $matches);
-            $descFunction = $this->sourceConnection->fetchAll(sprintf(
-                'DESC PROCEDURE %s.%s.%s(%s)',
-                Helper::quoteIdentifier($procedure['catalog_name']),
-                Helper::quoteIdentifier($procedure['schema_name']),
-                Helper::quoteIdentifier($procedure['name']),
-                $matches[1]
-            ));
-
-            $procedureParams = array_combine(
-                array_map(fn($v) => $v['property'], $descFunction),
-                array_map(fn($v) => $v['value'], $descFunction)
-            );
-
-            switch ($procedureParams['language']) {
-                case 'SQL':
-                    $procedureQuery = $this->buildSqlProcedureQuery($procedure, $procedureParams);
-                    break;
-                case 'JAVA':
-                    $procedureQuery = $this->buildJavaProcedureQuery($procedure, $procedureParams);
-                    break;
-                case 'JAVASCRIPT':
-                    $procedureQuery = $this->buildJavascriptProcedureQuery($procedure, $procedureParams);
-                    break;
-                case 'PYTHON':
-                    $procedureQuery = $this->buildPythonProcedureQuery($procedure, $procedureParams);
-                    break;
-                default:
-                    $this->logger->warning(sprintf(
-                        'Warning: Skip creating procedure "%s". Language "%s" is not supported.',
-                        Helper::quoteIdentifier($procedure['name']),
-                        $procedure['language']
-                    ));
-                    continue 2;
-            }
-
-            $ownership = array_filter(
-                $proceduresGrants,
-                fn($v) => str_contains($v['granted_by'], $procedure['schema_name'])
-            );
-
-            $this->destinationConnection->useRole(current($ownership)['granted_by']);
-
-            $this->destinationConnection->query(sprintf(
-                'USE SCHEMA %s.%s;',
-                Helper::quoteIdentifier($procedure['catalog_name']),
-                Helper::quoteIdentifier($procedure['schema_name'])
-            ));
-
-            try {
-                $this->destinationConnection->query($procedureQuery);
-            } catch (Throwable $e) {
-                $this->logger->info(sprintf(
-                    'Skip creating procedure %s.%s.%s. Error: "%s".',
-                    Helper::quoteIdentifier($procedure['catalog_name']),
-                    Helper::quoteIdentifier($procedure['schema_name']),
-                    Helper::quoteIdentifier($procedure['name']),
-                    $e->getMessage()
-                ));
-            }
-        }
-
-        foreach ($proceduresGrants as $proceduresGrant) {
-            if ($proceduresGrant['privilege'] === 'OWNERSHIP') {
-                continue;
-            }
-            $this->destinationConnection->assignGrantToRole($proceduresGrant);
-        }
-    }
-
-    private function buildSqlProcedureQuery(array $procedure, array $procedureParams): string
-    {
-        $sql = <<<SQL
-CREATE %s PROCEDURE %s%s
-returns %s 
-language sql
-execute as %s
-AS 
-%s
-;
-SQL;
-
-        return sprintf(
-            $sql,
-            $procedure['is_secure'] === 'Y' ? 'SECURE' : '',
-            Helper::quoteIdentifier($procedure['name']),
-            $procedureParams['signature'],
-            $procedureParams['returns'],
-            $procedureParams['execute as'],
-            trim($procedureParams['body'])
-        );
-    }
-
-    private function buildJavaProcedureQuery(array $procedure, array $procedureParams): string
-    {
-        $sql = <<<SQL
-CREATE %s PROCEDURE %s%s
-returns %s 
-language java
-runtime_version = '%s'
-handler = '%s'
-execute as %s
-%s
-%s
-AS 
-$$
-%s
-$$
-;
-SQL;
-
-        return sprintf(
-            $sql,
-            $procedure['is_secure'] === 'Y' ? 'SECURE' : '',
-            Helper::quoteIdentifier($procedure['name']),
-            $procedureParams['signature'],
-            $procedureParams['returns'],
-            $procedureParams['runtime_version'],
-            $procedureParams['handler'],
-            $procedureParams['execute as'],
-            $procedureParams['packages'] !== '[]' ?
-                'PACKAGES = (' . substr($procedureParams['packages'], 1, -1) . ')' :
-                '',
-            $procedureParams['imports'] !== '[]' ?
-                'IMPORTS = (' . substr($procedureParams['imports'], 1, -1) . ')' :
-                '',
-            trim($procedureParams['body'])
-        );
-    }
-
-    private function buildJavascriptProcedureQuery(array $procedure, array $procedureParams): string
-    {
-        $sql = <<<SQL
-CREATE %s PROCEDURE %s%s
-returns %s 
-language javascript
-execute as %s
-AS 
-$$
-%s
-$$
-;
-SQL;
-
-        return sprintf(
-            $sql,
-            $procedure['is_secure'] === 'Y' ? 'SECURE' : '',
-            Helper::quoteIdentifier($procedure['name']),
-            $procedureParams['signature'],
-            $procedureParams['returns'],
-            $procedureParams['execute as'],
-            trim($procedureParams['body'])
-        );
-    }
-
-    private function buildPythonProcedureQuery(array $procedure, array $procedureParams): string
-    {
-        $sql = <<<SQL
-CREATE %s PROCEDURE %s%s
-returns %s 
-language python
-runtime_version = '%s'
-handler = '%s'
-%s
-%s
-execute as %s
-AS 
-$$
-%s
-$$
-;
-SQL;
-
-        return sprintf(
-            $sql,
-            $procedure['is_secure'] === 'Y' ? 'SECURE' : '',
-            Helper::quoteIdentifier($procedure['name']),
-            $procedureParams['signature'],
-            $procedureParams['returns'],
-            $procedureParams['runtime_version'],
-            $procedureParams['handler'],
-            $procedureParams['packages'] !== '[]' ?
-                'PACKAGES = (' . substr($procedureParams['packages'], 1, -1) . ')' :
-                '',
-            $procedureParams['imports'] !== '[]' ?
-                'IMPORTS = (' . substr($procedureParams['imports'], 1, -1) . ')' :
-                '',
-            $procedureParams['execute as'],
-            trim($procedureParams['body'])
-        );
-    }
-}
